fix(utils.lua): fix blink completion item selection
The `M.cmp_select_nth` function now accepts the `entries` table as an argument.
This allows the function to be called with a specific set of entries, rather
than always using the entries from the last completion.

feat(utils.lua): add number option to get_rime_entry_ids

This commit adds a `number` option to the `get_rime_entry_ids` function.
This option allows the function to return the ID of a specific entry in the
list of entries.

fix(utils.lua): fix is_rime_entry function

The `is_rime_entry` function was not correctly identifying Rime entries when
using Blink. This commit fixes the issue by ensuring that the input code is
compared to the last characters of the result.