refactor(utils.lua): remove unnecessary stopinsert and feedkey calls from start_rime_ls
The `stopinsert` command is not needed as the function is not intended to exit insert mode.
The `M.feedkey("a", "n")` call was likely an artifact from previous development and is not necessary for the current functionality of `start_rime_ls`. Removing it cleans up the code and prevents unintended behavior.